created payload model class in place of locationHandler, and fixed firebase message service to only retrieve and broadcast payload message as opposed to working with data outside of application. Main activity for now will handle filtering
