--- app/src/main/java/edu/temple/contacttracer/LocationHandler.java
@@ -1,9 +1,0 @@
-package edu.temple.contacttracer;
-
-import android.location.LocationListener;
-import android.location.LocationManager;
-
-public class LocationHandler {
-    LocationManager locationManager;
-    LocationListener locationListener;
-}
