+++ app/src/main/java/edu/temple/contacttracer/MainActivity.java
@@ -2,14 +2,20 @@ package edu.temple.contacttracer;
 
 import android.Manifest;
 import android.app.ActivityManager;
+import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.pm.PackageManager;
+import android.location.Location;
+import android.location.LocationListener;
+import android.location.LocationManager;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import androidx.fragment.app.FragmentManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import androidx.room.Room;
 import androidx.room.RoomDatabase;
 import com.google.android.gms.tasks.OnCompleteListener;
@@ -17,6 +23,8 @@ import com.google.android.gms.tasks.Task;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 import edu.temple.contacttracer.database.AppDatabase;
 import edu.temple.contacttracer.database.ContactUUIDDao;
 import edu.temple.contacttracer.database.ContactUUIDModel;
@@ -33,7 +41,7 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
     DashboardFragment dashboardFragment;
 
     ArrayList<ContactUUIDModel> contactUUIDModelArrayList;
-    
+
     private static final int DB_SIZE_LIMIT = 14;
 
     private String payload;
@@ -42,15 +50,30 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
      * Check if application has GPS permissions, if not request them
      * add dashboardFragment to activity
      * updateUUID and log the current database
+     *
      * @param savedInstanceState
      */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        locationManager = getSystemService(LocationManager.class);
+
+        locationListener = new LocationListener() {
+            @Override
+            public void onLocationChanged(@NonNull Location location) {
+                currentLocation = location;
+            }
+        };
+
         if (!hasGPSPermission()) {
             requestGPSPermission();
         }
+        locationManager.requestLocationUpdates(LocationManager.GPS_PROVIDER, 0, 1, locationListener);
+        locationManager.requestLocationUpdates(LocationManager.NETWORK_PROVIDER, 0, 1, locationListener);
+        locationManager.requestLocationUpdates(LocationManager.PASSIVE_PROVIDER, 0, 1, locationListener);
+
         fm = getSupportFragmentManager();
 
         dashboardFragment = DashboardFragment.newInstance();
@@ -63,7 +86,7 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
         updateUUID();
         logDatabase();
 
-
+        LocalBroadcastManager.getInstance(this).registerReceiver(broadcastReceiver, new IntentFilter("FMS"));
 
     }
 
@@ -127,11 +150,11 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
 
     }
 
-    public static List<ContactUUIDModel> getContactModelList(Context context){
+    public static List<ContactUUIDModel> getContactModelList(Context context) {
         AppDatabase db = Room.databaseBuilder(context, AppDatabase.class, "uuid-database").build();
         final ContactUUIDDao contactUUIDDao = db.contactUUIDDao();
         final List<ContactUUIDModel>[] contactUUIDModelList = new List[]{new ArrayList<>()};
-        Thread thread = new Thread(){
+        Thread thread = new Thread() {
 
             @Override
             public void run() {
@@ -152,19 +175,19 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
     /**
      * Reset Database back to size 0 for testing
      */
-    public void resetDatabase(){
+    public void resetDatabase() {
         AppDatabase db = Room.databaseBuilder(getApplicationContext(), AppDatabase.class, "uuid-database").build();
         final ContactUUIDDao contactUUIDDao = db.contactUUIDDao();
 
-        Thread thread = new Thread(){
+        Thread thread = new Thread() {
 
 
             @Override
             public void run() {
                 contactUUIDModelArrayList = (ArrayList<ContactUUIDModel>) contactUUIDDao.getAll();
-                if (contactUUIDModelArrayList != null){
-                    if (contactUUIDModelArrayList.size() > 0){
-                        for (ContactUUIDModel contactUUIDModel: contactUUIDModelArrayList){
+                if (contactUUIDModelArrayList != null) {
+                    if (contactUUIDModelArrayList.size() > 0) {
+                        for (ContactUUIDModel contactUUIDModel : contactUUIDModelArrayList) {
                             contactUUIDDao.delete(contactUUIDModel.index);
                         }
                     }
@@ -185,18 +208,18 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
     /**
      * Log all entries of database
      */
-    public void logDatabase(){
+    public void logDatabase() {
 
         AppDatabase db = Room.databaseBuilder(getApplicationContext(), AppDatabase.class, "uuid-database").build();
         final ContactUUIDDao contactUUIDDao = db.contactUUIDDao();
 
-        Thread thread = new Thread(){
+        Thread thread = new Thread() {
             @Override
             public void run() {
                 String database = "";
                 ArrayList<ContactUUIDModel> contactUUIDModelArrayList = (ArrayList<ContactUUIDModel>) contactUUIDDao.getAll();
 
-                for (ContactUUIDModel contactUUIDModel : contactUUIDModelArrayList){
+                for (ContactUUIDModel contactUUIDModel : contactUUIDModelArrayList) {
                     database += contactUUIDModel.toString() + "\n";
                 }
                 Log.i("logdb", "logDatabase: \n" + database);
@@ -213,7 +236,6 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
     }
 
     /**
-     *
      * @param serviceClass service to be checked
      * @return true if service is running, false if service is not running
      */
@@ -228,7 +250,6 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
     }
 
     /**
-     *
      * @return if service has or does not have GPS permission
      */
     private boolean hasGPSPermission() {
@@ -261,4 +282,91 @@ public class MainActivity extends AppCompatActivity implements DashboardFragment
         stopService(serviceIntent);
         Log.i("LocatorService", "Stopping LocatorService ");
     }
+
+
+    private BroadcastReceiver broadcastReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            payload = intent.getStringExtra("payload");
+            boolean isFilteredPayload = isGrater6Feet_differentUUID(payload);
+            Log.i("filter", "onReceive: " + isFilteredPayload);
+        }
+    };
+
+    @Override
+    protected void onDestroy() {
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(broadcastReceiver);
+        super.onDestroy();
+    }
+
+    LocationManager locationManager;
+    LocationListener locationListener;
+    Location currentLocation;
+
+    public void storePayload(PayloadModel payloadModel){
+        
+    }
+
+    public boolean isGrater6Feet_differentUUID(String payload) {
+
+        final PayloadModel payloadModel = generatePayloadModel(payload);
+
+        Location location = new Location("");
+        double latitude = Double.parseDouble(payloadModel.getLatitude());
+        double longitude = Double.parseDouble(payloadModel.getLongitude());
+        Log.i("locationCheck", "latitude: " + currentLocation.getLatitude() + "\nlongitude: " + currentLocation.getLongitude());
+        location.setLatitude(latitude);
+        location.setLongitude(longitude);
+
+        // check if database contains uuid
+        final boolean[] containsUUID = {true};
+        Thread thread = new Thread() {
+            @Override
+            public void run() {
+                super.run();
+                List<ContactUUIDModel> contactUUIDModelList = MainActivity.getContactModelList(getApplicationContext());
+                assert contactUUIDModelList != null;
+                for (ContactUUIDModel model : contactUUIDModelList) {
+                    if (model.uuid.equals(payloadModel.getUuid())) {
+                        break;
+                    }
+                    containsUUID[0] = false;
+                }
+            }
+        };
+
+        thread.start();
+        try {
+            thread.join();
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+            return false;
+        }
+
+        // check distance between current location and payload location
+        float[] results = new float[1];
+        Log.i("Filter", "containsUUID: " + Arrays.toString(containsUUID) + " distance: " + results[0]);
+        Location.distanceBetween(latitude, longitude, currentLocation.getLatitude(), currentLocation.getLongitude(), results);
+
+        return results[0] >= 1.8288 && containsUUID[0] == false;
+    }
+
+    private PayloadModel generatePayloadModel(String payload){
+
+        try {
+            JsonObject payloadObject = new JsonParser().parse(payload).getAsJsonObject();
+            String uuid = payloadObject.get("uuid").getAsString();
+            String latitude = payloadObject.get("latitude").getAsString();
+            String longitude = payloadObject.get("longitude").getAsString();
+            String sedentary_begin = payloadObject.get("sedentary_begin").getAsString();
+            String sedentary_end = payloadObject.get("sedentary_end").getAsString();
+            PayloadModel model = new PayloadModel(uuid, latitude, longitude, sedentary_begin, sedentary_end);
+            Log.i("PAYLOADMODEL", "generatePayloadModel: " + model.toString());
+            return model;
+        } catch(Exception e){
+            e.printStackTrace();
+            return null;
+        }
+    }
 }
+
